# apps/api/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MarkNotificationsReadRequest": {
        "properties": {
          "ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 50,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ids"
          }
        },
        "title": "MarkNotificationsReadRequest",
        "type": "object"
      },
      "MarkNotificationsReadResponse": {
        "properties": {
          "status": {
            "const": "ok",
            "title": "Status",
            "type": "string"
          },
          "unreadCount": {
            "title": "Unreadcount",
            "type": "integer"
          }
        },
        "required": [
          "status",
          "unreadCount"
        ],
        "title": "MarkNotificationsReadResponse",
        "type": "object"
      },
...
      "NotificationActor": {
        "properties": {
          "avatarUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatarurl"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "NotificationActor",
        "type": "object"
      },
      "NotificationItem": {
        "properties": {
          "actor": {
            "$ref": "#/components/schemas/NotificationActor"
          },
          "commentText": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commenttext"
          },
          "cookedNote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cookednote"
          },
          "cookedRating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cookedrating"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "post": {
            "$ref": "#/components/schemas/NotificationPost"
          },
          "reactionSummary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReactionSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "readAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Readat"
          },
          "type": {
            "enum": [
              "comment",
              "reaction_batch",
              "cooked"
            ],
            "title": "Type",
            "type": "string"
          },
          "updatedAt": {
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "createdAt",
          "updatedAt",
          "actor",
          "post"
        ],
        "title": "NotificationItem",
        "type": "object"
      },
      "NotificationPost": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "mainPhotoUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mainphotourl"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title"
        ],
        "title": "NotificationPost",
        "type": "object"
      },
      "NotificationsListResponse": {
        "properties": {
          "hasMore": {
            "title": "Hasmore",
            "type": "boolean"
          },
          "nextOffset": {
            "title": "Nextoffset",
            "type": "integer"
          },
          "notifications": {
            "items": {
              "$ref": "#/components/schemas/NotificationItem"
            },
            "title": "Notifications",
            "type": "array"
          },
          "unreadCount": {
            "title": "Unreadcount",
            "type": "integer"
          }
        },
        "required": [
          "notifications",
          "unreadCount",
          "hasMore",
          "nextOffset"
        ],
        "title": "NotificationsListResponse",
        "type": "object"
      },
      "ReactionEmojiCount": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "emoji": {
            "title": "Emoji",
            "type": "string"
          }
        },
        "required": [
          "emoji",
          "count"
        ],
        "title": "ReactionEmojiCount",
        "type": "object"
      },
...
      "ReactionSummary": {
        "properties": {
          "emojiCounts": {
            "items": {
              "$ref": "#/components/schemas/ReactionEmojiCount"
            },
            "title": "Emojicounts",
            "type": "array"
          },
          "lastEmoji": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastemoji"
          },
          "totalCount": {
            "title": "Totalcount",
            "type": "integer"
          }
        },
        "required": [
          "totalCount",
          "emojiCounts"
        ],
        "title": "ReactionSummary",
        "type": "object"
      },
...
      "UnreadCountResponse": {
        "properties": {
          "unreadCount": {
            "title": "Unreadcount",
            "type": "integer"
          }
        },
        "required": [
          "unreadCount"
        ],
        "title": "UnreadCountResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Family Recipe API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/timeline": {
      "get": {
        "operationId": "get_timeline_timeline_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Timeline",
        "tags": [
          "timeline"
        ]
      }
    },
...
    "/v1/notifications": {
      "get": {
        "operationId": "list_notifications_v1_notifications_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "unreadOnly",
            "required": false,
            "schema": {
              "default": false,
              "title": "Unreadonly",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationsListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Notifications",
        "tags": [
          "notifications"
        ]
      }
    },
    "/v1/notifications/mark-read": {
      "post": {
        "operationId": "mark_read_v1_notifications_mark_read_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkNotificationsReadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkNotificationsReadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mark Read",
        "tags": [
          "notifications"
        ]
      }
    },
    "/v1/notifications/unread-count": {
      "get": {
        "operationId": "unread_count_v1_notifications_unread_count_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnreadCountResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Unread Count",
        "tags": [
          "notifications"
        ]
      }
    },
...
    "/v1/timeline": {
      "get": {
        "operationId": "get_timeline_v1_timeline_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Timeline",
        "tags": [
          "timeline"
        ]
      }
    }
  }
}